feat: `implement POST, GET and DELETE routes` for course management. - Added a DELETE route to handle the deletion of a course. - Added a POST route for creating a new course. - Added a GET route for retrieving all courses.

+++ app/api/courses/[courseId]/route.ts
@@ -1,5 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { db as Prisma } from "@/lib/db";
+import { auth } from "@clerk/nextjs";
+import { isTeacher } from "@/lib/isTeacher";
 
 export async function DELETE(
   request: NextRequest,
@@ -8,15 +10,22 @@ export async function DELETE(
   const courseId = params.courseId;
   if (courseId) {
     try {
+      const { userId } = auth();
+
+      if (!userId || !isTeacher(userId)) {
+        return NextResponse.json("Not Authorized", { status: 401 });
+      }
       const res = await Prisma.course.delete({
         where: {
           id: courseId,
         },
       });
-      return NextResponse.json(`Course ${res.title} successfuly deleted`);
+      return NextResponse.json(`Course ${res.title} successfuly deleted`, {
+        status: 201,
+      });
     } catch (error) {
       return NextResponse.json(`Unable to delete the course Error: ${error}`);
     }
   }
-  return NextResponse.json(`Course doens't exist`);
+  return NextResponse.json(`Unable to retrive data`, { status: 500 });
 }
